Major structural overhaul
Some notes that should’ve happened in smaller commits:

Turned Catphish into it’s own module.

Rely on trollop for command-line option parsing instead of whatever
getoptlong stuff.

Instead of calling the uniq on the array of domains, lets just use a
set to help get rid of duplicates while populating the domain
containers.

I’ve also moved some of the instance variables in the application to
constants in the Catphish module.

so, for example `@VERSION` can be accused via `Catphish::VERSION`

All of the methods for constructing domains can be accessed via the
Catphish module as well:

for example:

`Catphish.prepend_or_append(domain)`

We can loop over the potential methods by having then in options
available in trollop and then `.send` them as their symbol form to
Catfish with the domain as the second argument. This is will just call
the method on the under the hoof without dealing with too many nasty
if/when/each loops.

The logo has been moved to it’s own method.

The `.start` method is the “main” method of sorts of the application.

I also added my name as part of the author section.  Lemme know if you
don’t want that. ¯\_(ツ)_/¯