Create a dockerfile for easy, sandboxed execution of the application
Linux containers are a primitive for executing processes in a sandbox
environment through a combination of linux namespaces and cgroups.
Docker is the most popular implementation of this abstraction, and comes
with a reasonable behaviour in which it isolates a process with these
namespaces and with seccomp, but does not (by default) apply cgroup
limits. However, it offers a nice chroot style environment where the
application can be trivially spun up and executed.

This commit adds a Dockerfile, which will successfully build the
program and allow it to be executed. Instructions for executing it are
documented in comments in the Dockerfile itself, but should be familiar
to most regular docker users.