Swap the README line ending for unix style line endings, rather than Windows
The line endings on a file offer little semantic different to the user
in most cases and have little impact on the application. This is
particularly the case on the README file which is never consumed by any
application. However, generally speaking a project is expected to follow
a consistent line ending format. This project appears to use unix style
line endings in all cases except the README, and the `git` version
control tool complains of whitespace errors when committing the file.

This commit changes the README to use unix line endings. It further
removes any redundant whitespace from the end of the line.