Add some documentation on the use of Docker
Previous work had added a Dockerfile to the root of the repository which
allows the invocation of catphish, without the installation of ruby or
it's dependencies. However, for newer users or those unclear on the use
or purpose of this container there are no steps that explain it's use.

This commit adds documentation to the readme explaining it's purpose,
and documenting how to use it. The commands in the README should be copy
paste safe, and were tested on the authors Debian machine.