Create a changelog
The project has undergone several significant steps in it's growth,
culminating in the recent release of the 1.0 release. Hopefully it will
continue to do so. To track these changes over time, a document called a
"Changelog" is usually created, documenting what developers have added,
broken, or changed.

This commit adds such a document. The format supplied by
"keepachangelog.org" has been used, as it's been used by the author in
other projects, provides an impartial standard and seems reasonably
sane.